refactor(auth): move signup verification hook to feature hooks directory
The signup verification mutation hook was relocated from the api directory to the feature-specific hooks directory to better organize the codebase and follow feature-based architecture principles.
